wrc.h: fix cast of pointer to long, not int
Otherwise, the macro won't build on 64-bit hosts

+++ include/wrc.h
@@ -21,7 +21,7 @@
 #define sprintf pp_sprintf
 
 #undef offsetof
-#define offsetof(TYPE, MEMBER) ((int) &((TYPE *)0)->MEMBER)
+#define offsetof(TYPE, MEMBER) ((long) &((TYPE *)0)->MEMBER)
 #undef ARRAY_SIZE
 #define ARRAY_SIZE(arr) (sizeof(arr) / sizeof((arr)[0]))
